Assignment 5.3
Completed both cURL and Postman requests with code examples. I reverted the changes to the json file (POST request as stated in this doc) to preserve the to-do-service database from this pull request.

Removing page

Removing my-api-example page so the focus is the to-do-example for this branch.

Fixes after review

Markdownlint violations fixed